Change os2/Charset to basic-latin and ttf generating options(with DSIG and windows-compatible kern checked)
This issue was found when the font was used in Eclispe, and some Chinese
characters are displayed correct, some are not, which was really weird
and I have to compare every single settings between cerez.sfd and other
fonts which are good with eclipse such as cascadiax, consolasx., etc.
But there are not fundamental differences to effect this issue.
Then I have to check the generating options...